+++ app/models/panda/user.rb
@@ -1,4 +1,9 @@
 module Panda
   class User < ApplicationRecord
+      # Include default devise modules. Others available are:
+  # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
+    devise :database_authenticatable, :registerable,
+    :recoverable, :rememberable, :validatable
+    include DeviseTokenAuth::Concerns::User
   end
 end