+++ config/routes.rb
@@ -1,2 +1,5 @@
+require_relative 'routes/api'
+
 Panda::Engine.routes.draw do
+  resource :wechat, only: [:show, :create]
 end