+++ lib/panda.rb
@@ -1,3 +1,10 @@
+# requires all dependencies
+Gem.loaded_specs['panda'].dependencies.each do |d|
+  require d.name
+ end
+
+require 'devise'
+
 require "panda/version"
 require "panda/engine"