+++ test/dummy/db/schema.rb
@@ -10,7 +10,34 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_06_22_155535) do
+ActiveRecord::Schema.define(version: 2021_06_25_015354) do
+
+  create_table "panda_identities", force: :cascade do |t|
+    t.integer "user_id", null: false
+    t.string "uid"
+    t.string "provider"
+    t.string "status"
+    t.string "token"
+    t.string "refresh_token"
+    t.string "scopes"
+    t.integer "expires_at"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.index ["provider"], name: "index_panda_identities_on_provider"
+    t.index ["refresh_token"], name: "index_panda_identities_on_refresh_token"
+    t.index ["status"], name: "index_panda_identities_on_status"
+    t.index ["token"], name: "index_panda_identities_on_token"
+    t.index ["uid"], name: "index_panda_identities_on_uid"
+    t.index ["user_id"], name: "index_panda_identities_on_user_id"
+  end
+
+  create_table "panda_profiles", force: :cascade do |t|
+    t.string "store"
+    t.integer "user_id", null: false
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.index ["user_id"], name: "index_panda_profiles_on_user_id"
+  end
 
   create_table "panda_users", force: :cascade do |t|
     t.string "name"
@@ -22,4 +49,6 @@ ActiveRecord::Schema.define(version: 2021_06_22_155535) do
     t.datetime "updated_at", precision: 6, null: false
   end
 
+  add_foreign_key "panda_identities", "users"
+  add_foreign_key "panda_profiles", "users"
 end