+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 
+# 2021.06.25
+** Features **
++ Model for user
++ Model for identity
++ Model for profile
++ Login api with devise token
++ Login via wechat